Add new metric for Quiet Mode status (#687)
* New metric indicating whether Jenkins is in quietdown mode

This metric exposes the Jenkins isQuietingDown status as a gauge.
When Jenkins is in quiet mode, the new metric will represent this
as a '1', otherwise the value will be '0'

Exposing this metric allows Prometheus to determine whether Jenkins
is in quiet mode.  This can be useful in cases where an alert should
be triggered if quiet mode has been enabled for too long or falls
outside of a maintenance window.

Fixes issue #686

* Fixing JenkinsStatusCollectorTest unit test

The unit test checks the number of status metrics.
Adding a new metric caused this test to fail.
The number of metrics needed to be incremented to fix the test.